Development 4.1.5-beta.1

+++ OpenBCI_GUI/SoftwareSettings.pde
@@ -379,12 +379,12 @@ class SoftwareSettings {
             this.load(settingsFileToLoad);
             errorUserSettingsNotFound = false;
         } catch (Exception e) {
-            e.printStackTrace();
+            //e.printStackTrace();
             println("InitSettings: " + settingsFileToLoad + " not found or other error.");
             errorUserSettingsNotFound = true;
-            File f = new File(sketchPath()+System.getProperty("file.separator")+settingsFileToLoad);
+            File f = new File(settingsFileToLoad);
             if (f.exists()) {
-                if (f.delete()) println("SoftwareSettings: Removed old settings file.");
+                if (f.delete()) println("SoftwareSettings: Removed old/broken settings file.");
             }
         }
     }
@@ -450,7 +450,7 @@ class SoftwareSettings {
                     saveTSSettingsJSONArray.setJSONObject(i, saveChannelSettings);
                 } //end channel settings for loop
             } //end all channels for loop
-            saveSettingsJSONData.setJSONArray("channelSettings", saveTSSettingsJSONArray); //Set the JSON array for all channels
+            saveTSSettings.setJSONArray("channelSettings", saveTSSettingsJSONArray); //Set the JSON array for all channels
         }
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
         //              Case for saving TS settings when in Ganglion, Synthetic, and Playback data modes                       //
@@ -1239,18 +1239,18 @@ class SoftwareSettings {
                 w_networking.cp5_networking_dropdowns.get(ScrollableList.class, "dataType3").setValue(nwDataType3);
                 w_networking.cp5_networking_dropdowns.getController("dataType4").getCaptionLabel().setText(nwDataTypesArray[nwDataType4]);
                 w_networking.cp5_networking_dropdowns.get(ScrollableList.class, "dataType4").setValue(nwDataType4);
-                w_networking.cp5_networking.get(Textfield.class, "osc_ip1").setText(nwOscIp1Load); //Simply set the text for text boxes
-                w_networking.cp5_networking.get(Textfield.class, "osc_ip2").setText(nwOscIp2Load); //The strings are referenced on command
-                w_networking.cp5_networking.get(Textfield.class, "osc_ip3").setText(nwOscIp3Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_ip4").setText(nwOscIp4Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_port1").setText(nwOscPort1Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_port2").setText(nwOscPort2Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_port3").setText(nwOscPort3Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_port4").setText(nwOscPort4Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_address1").setText(nwOscAddress1Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_address2").setText(nwOscAddress2Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_address3").setText(nwOscAddress3Load);
-                w_networking.cp5_networking.get(Textfield.class, "osc_address4").setText(nwOscAddress4Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_ip1").setText(nwOscIp1Load); //Simply set the text for text boxes
+                w_networking.cp5_networking.get(Textfield.class, "OSC_ip2").setText(nwOscIp2Load); //The strings are referenced on command
+                w_networking.cp5_networking.get(Textfield.class, "OSC_ip3").setText(nwOscIp3Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_ip4").setText(nwOscIp4Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_port1").setText(nwOscPort1Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_port2").setText(nwOscPort2Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_port3").setText(nwOscPort3Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_port4").setText(nwOscPort4Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_address1").setText(nwOscAddress1Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_address2").setText(nwOscAddress2Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_address3").setText(nwOscAddress3Load);
+                w_networking.cp5_networking.get(Textfield.class, "OSC_address4").setText(nwOscAddress4Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter1").activate(nwOscFilter1Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter2").activate(nwOscFilter2Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter3").activate(nwOscFilter3Load);
@@ -1264,12 +1264,12 @@ class SoftwareSettings {
                 w_networking.cp5_networking_dropdowns.get(ScrollableList.class, "dataType2").setValue(nwDataType2);
                 w_networking.cp5_networking_dropdowns.getController("dataType3").getCaptionLabel().setText(nwDataTypesArray[nwDataType3]);
                 w_networking.cp5_networking_dropdowns.get(ScrollableList.class, "dataType3").setValue(nwDataType3);
-                w_networking.cp5_networking.get(Textfield.class, "udp_ip1").setText(nwUdpIp1Load);
-                w_networking.cp5_networking.get(Textfield.class, "udp_ip2").setText(nwUdpIp2Load);
-                w_networking.cp5_networking.get(Textfield.class, "udp_ip3").setText(nwUdpIp3Load);
-                w_networking.cp5_networking.get(Textfield.class, "udp_port1").setText(nwUdpPort1Load);
-                w_networking.cp5_networking.get(Textfield.class, "udp_port2").setText(nwUdpPort2Load);
-                w_networking.cp5_networking.get(Textfield.class, "udp_port3").setText(nwUdpPort3Load);
+                w_networking.cp5_networking.get(Textfield.class, "UDP_ip1").setText(nwUdpIp1Load);
+                w_networking.cp5_networking.get(Textfield.class, "UDP_ip2").setText(nwUdpIp2Load);
+                w_networking.cp5_networking.get(Textfield.class, "UDP_ip3").setText(nwUdpIp3Load);
+                w_networking.cp5_networking.get(Textfield.class, "UDP_port1").setText(nwUdpPort1Load);
+                w_networking.cp5_networking.get(Textfield.class, "UDP_port2").setText(nwUdpPort2Load);
+                w_networking.cp5_networking.get(Textfield.class, "UDP_port3").setText(nwUdpPort3Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter1").activate(nwUdpFilter1Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter2").activate(nwUdpFilter2Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter3").activate(nwUdpFilter3Load);
@@ -1282,15 +1282,15 @@ class SoftwareSettings {
                 w_networking.cp5_networking_dropdowns.get(ScrollableList.class, "dataType2").setValue(nwDataType2);
                 w_networking.cp5_networking_dropdowns.getController("dataType3").getCaptionLabel().setText(nwDataTypesArray[nwDataType3]);
                 w_networking.cp5_networking_dropdowns.get(ScrollableList.class, "dataType3").setValue(nwDataType3);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_name1").setText(nwLSLName1Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_name2").setText(nwLSLName2Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_name3").setText(nwLSLName3Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_type1").setText(nwLSLType1Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_type2").setText(nwLSLType2Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_type3").setText(nwLSLType3Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_numchan1").setText(nwLSLNumChan1Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_numchan2").setText(nwLSLNumChan2Load);
-                w_networking.cp5_networking.get(Textfield.class, "lsl_numchan3").setText(nwLSLNumChan3Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_name1").setText(nwLSLName1Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_name2").setText(nwLSLName2Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_name3").setText(nwLSLName3Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_type1").setText(nwLSLType1Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_type2").setText(nwLSLType2Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_type3").setText(nwLSLType3Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_numchan1").setText(nwLSLNumChan1Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_numchan2").setText(nwLSLNumChan2Load);
+                w_networking.cp5_networking.get(Textfield.class, "LSL_numchan3").setText(nwLSLNumChan3Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter1").activate(nwLSLFilter1Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter2").activate(nwLSLFilter2Load);
                 w_networking.cp5_networking.get(RadioButton.class, "filter3").activate(nwLSLFilter3Load);
@@ -1532,7 +1532,7 @@ class SoftwareSettings {
             output("The new data source is " + dataModeVersionToPrint + " and NCHAN = [" + nchan + "]. Data source error: Default Settings Loaded."); //Show a normal message for loading Default Settings
         } else if (errorUserSettingsNotFound) {
             verbosePrint("OpenBCI_GUI: initSystem: -- Init 5 -- " + "Load settings error: File not found. " + millis()); //Print the error to console
-            output("The new data source is " + dataModeVersionToPrint + " and NCHAN = [" + nchan + "]. User settings error: Default Settings Loaded."); //Show a normal message for loading Default Settings
+            output("The new data source is " + dataModeVersionToPrint + " and NCHAN = [" + nchan + "]. User Settings Error: Default Settings Loaded."); //Show a normal message for loading Default Settings
         } else {
             verbosePrint("OpenBCI_GUI: initSystem: -- Init 5 -- " + "Load settings error: Connection Error: Failed to apply channel settings to Cyton" + millis()); //Print the error to console
             outputError(dataModeVersionToPrint + " and NCHAN = [" + nchan + "]. Connection Error: Channel settings failed to apply to Cyton."); //Show a normal message for loading Default Settings
@@ -1553,14 +1553,27 @@ class SoftwareSettings {
             errorUserSettingsNotFound = true;
         }
         //Output message when Loading settings is complete
+        String err = null;
         if (chanNumError == false && dataSourceError == false && errorUserSettingsNotFound == false && loadErrorCytonEvent == false) {
             outputSuccess("Settings Loaded!");
         } else if (chanNumError) {
-            outputError("Load Settings Error: Invalid number of channels");
+            err = "Invalid number of channels";
         } else if (dataSourceError) {
-            outputError("Load Settings Error: Invalid data source");
-        } else {
-            outputError("Load settings error: " + settingsFileToLoad + " not found. ");
+            err = "Invalid data source";
+        } else if (errorUserSettingsNotFound) {
+            err = settingsFileToLoad + " not found.";
+        }
+
+        if (err != null) {
+            println("Load Settings Error: " + err);
+            File f = new File(settingsFileToLoad);
+            if (f.exists()) {
+                if (f.delete()) {
+                    outputError("Found old/broken GUI settings. Please reconfigure the GUI and save new settings.");
+                } else {
+                    outputError("SoftwareSettings: Error deleting old/broken settings file...");
+                }
+            }
         }
     }
 
@@ -1595,14 +1608,15 @@ class SoftwareSettings {
             this.load(defaultSettingsFileToLoad);
             outputSuccess("Default Settings Loaded!");
         } catch (Exception e) {
-            outputError("Default Settings Error: New settings will be made next system start.");
-            File f = new File(sketchPath()+System.getProperty("file.separator")+defaultSettingsFileToLoad);
-            if (f.delete()) {
-                println("Old settings file succesfully deleted.");
-            } else {
-                println("Error deleting file...");
+            outputError("Default Settings Error: Valid Default Settings will be saved next system start.");
+            File f = new File(defaultSettingsFileToLoad);
+            if (f.exists()) {
+                if (f.delete()) {
+                    println("SoftwareSettings: Old/Broken Default Settings file succesfully deleted.");
+                } else {
+                    println("SoftwareSettings: Error deleting Default Settings file...");
+                }
             }
-
         }
     }
 
@@ -1659,8 +1673,8 @@ void loadConfigFile(File selection) {
                 outputError("Settings Error: Data Source Mismatch Detected");
             } else {
                 outputError("Error trying to load settings file, possibly from previous GUI. Removing old settings.");
+                if (selection.exists()) selection.delete();
             }
-            if (selection.exists()) selection.delete();
         }
         settings.loadDialogName = null; //reset this variable for future use
     }
